@@ -3,89 +3,91 @@ import 'package:core/data/data.dart';
 import 'package:equatable/equatable.dart';
 
 class TVDetailResponse extends Equatable {
- final bool adult;
- final String? backdropPath;
- final List<int> episodeRunTime;
- final String firstAirDate;
- final List<GenreModel> genres;
- final String homepage;
- final int id;
- final bool inProduction;
- final String? lastAirDate;
- final LastEpisodeToAirModel lastEpisodeToAir;
- final String name;
- final int numberOfEpisodes;
- final int numberOfSeasons;
- final List<String> originCountry;
- final String originalLanguage;
- final String originalName;
- final String overview;
- final double popularity;
- final String? posterPath;
- final List<SeasonsModel> seasons;
- final String status;
- final String tagline;
- final String type;
- final double voteAverage;
- final int voteCount;
+  final bool adult;
+  final String? backdropPath;
+  final List<int> episodeRunTime;
+  final String firstAirDate;
+  final List<GenreModel> genres;
+  final String homepage;
+  final int id;
+  final bool inProduction;
+  final String? lastAirDate;
+  final LastEpisodeToAirModel lastEpisodeToAir;
+  final String name;
+  final int numberOfEpisodes;
+  final int numberOfSeasons;
+  final List<String> originCountry;
+  final String originalLanguage;
+  final String originalName;
+  final String overview;
+  final double popularity;
+  final String? posterPath;
+  final List<SeasonsModel> seasons;
+  final String status;
+  final String tagline;
+  final String type;
+  final double voteAverage;
+  final int voteCount;
 
-  const TVDetailResponse(
-      {
-        required this.adult,
-        required this.backdropPath,
-        required this.episodeRunTime,
-        required this.firstAirDate,
-        required this.genres,
-        required this.homepage,
-        required this.id,
-        required this.inProduction,
-        required this.lastAirDate,
-        required this.lastEpisodeToAir,
-        required this.name,
-        required this.numberOfEpisodes,
-        required this.numberOfSeasons,
-        required this.originCountry,
-        required this.originalLanguage,
-        required this.originalName,
-        required this.overview,
-        required this.popularity,
-        required this.posterPath,
-        required this.seasons,
-        required this.status,
-        required this.tagline,
-        required this.type,
-        required this.voteAverage,
-        required this.voteCount,
-      });
+  const TVDetailResponse({
+    required this.adult,
+    required this.backdropPath,
+    required this.episodeRunTime,
+    required this.firstAirDate,
+    required this.genres,
+    required this.homepage,
+    required this.id,
+    required this.inProduction,
+    required this.lastAirDate,
+    required this.lastEpisodeToAir,
+    required this.name,
+    required this.numberOfEpisodes,
+    required this.numberOfSeasons,
+    required this.originCountry,
+    required this.originalLanguage,
+    required this.originalName,
+    required this.overview,
+    required this.popularity,
+    required this.posterPath,
+    required this.seasons,
+    required this.status,
+    required this.tagline,
+    required this.type,
+    required this.voteAverage,
+    required this.voteCount,
+  });
 
-  factory TVDetailResponse.fromJson(Map<String, dynamic> json) => TVDetailResponse(
-      adult: json["adult"],
-      backdropPath: json["backdrop_path"],
-      episodeRunTime: json["episode_run_time"].cast<int>(),
-      firstAirDate: json["first_air_date"],
-      genres: List<GenreModel>.from(
-          json["genres"].map((x) => GenreModel.fromJson(x))),
-      homepage: json["homepage"],
-      id: json["id"],
-      inProduction: json["in_production"],
-      lastAirDate: json["last_air_date"],
-      lastEpisodeToAir: LastEpisodeToAirModel.fromJson(json["last_episode_to_air"]),
-      name: json["name"],
-      numberOfEpisodes: json["number_of_episodes"],
-      numberOfSeasons: json["number_of_seasons"],
-      originCountry: json["origin_country"].cast<String>(),
-      originalLanguage: json["original_language"],
-      originalName: json["original_name"],
-      overview: json["overview"],
-      popularity: json["popularity"],
-      posterPath: json["poster_path"],
-      seasons: List<SeasonsModel>.from(json["seasons"].map((x) => SeasonsModel.fromJson(x))),
-      status: json["status"],
-      tagline: json["tagline"],
-      type: json["type"],
-      voteAverage: json["vote_average"],
-      voteCount: json["vote_count"],
-  );
+  factory TVDetailResponse.fromJson(Map<String, dynamic> json) =>
+      TVDetailResponse(
+        adult: json["adult"],
+        backdropPath: json["backdrop_path"],
+        episodeRunTime: json["episode_run_time"].cast<int>(),
+        firstAirDate: json["first_air_date"],
+        genres: List<GenreModel>.from(
+            json["genres"].map((x) => GenreModel.fromJson(x))),
+        homepage: json["homepage"],
+        id: json["id"],
+        inProduction: json["in_production"],
+        lastAirDate: json["last_air_date"],
+        lastEpisodeToAir:
+            LastEpisodeToAirModel.fromJson(json["last_episode_to_air"]),
+        name: json["name"],
+        numberOfEpisodes: json["number_of_episodes"],
+        numberOfSeasons: json["number_of_seasons"],
+        originCountry: json["origin_country"].cast<String>(),
+        originalLanguage: json["original_language"],
+        originalName: json["original_name"],
+        overview: json["overview"],
+        popularity: json["popularity"],
+        posterPath: json["poster_path"],
+        seasons: List<SeasonsModel>.from(
+            json["seasons"].map((x) => SeasonsModel.fromJson(x))),
+        status: json["status"],
+        tagline: json["tagline"],
+        type: json["type"],
+        voteAverage: json["vote_average"],
+        voteCount: json["vote_count"],
+      );
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = <String, dynamic>{};
@@ -149,29 +151,29 @@ class TVDetailResponse extends Equatable {
 
   @override
   List<Object?> get props => [
-    adult,
-    backdropPath,
-    firstAirDate,
-    genres,
-    homepage,
-    id,
-    inProduction,
-    lastAirDate,
-    lastEpisodeToAir,
-    name,
-    numberOfEpisodes,
-    numberOfSeasons,
-    originCountry,
-    originalLanguage,
-    originalName,
-    overview,
-    popularity,
-    posterPath,
-    seasons,
-    status,
-    tagline,
-    type,
-    voteAverage,
-    voteCount,
-  ];
+        adult,
+        backdropPath,
+        firstAirDate,
+        genres,
+        homepage,
+        id,
+        inProduction,
+        lastAirDate,
+        lastEpisodeToAir,
+        name,
+        numberOfEpisodes,
+        numberOfSeasons,
+        originCountry,
+        originalLanguage,
+        originalName,
+        overview,
+        popularity,
+        posterPath,
+        seasons,
+        status,
+        tagline,
+        type,
+        voteAverage,
+        voteCount,
+      ];
 }
