@@ -54,16 +54,16 @@ class MovieLocalDataSourceImpl implements MovieLocalDataSource {
   @override
   Future<void> cacheNowPlayingMovies(List<MovieTable> movies) async {
     await databaseHelper.clearCache('now playing');
-    await databaseHelper.insertCacheTransaction(movies,'now playing');
+    await databaseHelper.insertCacheTransaction(movies, 'now playing');
   }
 
   @override
   Future<List<MovieTable>> getCachedNowPlayingMovies() async {
-   final result = await databaseHelper.getCacheMovies('now playing');
-   if (result.isNotEmpty) {
-     return result.map((data) => MovieTable.fromMap(data)).toList();
-   } else {
-    throw CacheException("Can't get the data :(");
-   }
+    final result = await databaseHelper.getCacheMovies('now playing');
+    if (result.isNotEmpty) {
+      return result.map((data) => MovieTable.fromMap(data)).toList();
+    } else {
+      throw CacheException("Can't get the data :(");
+    }
   }
 }
