@@ -77,16 +77,16 @@ class LastEpisodeToAirModel extends Equatable {
 
   @override
   List<Object?> get props => [
-    airDate,
-    episodeNumber,
-    id,
-    name,
-    overview,
-    productionCode,
-    seasonNumber,
-    showId,
-    stillPath,
-    voteAverage,
-    voteCount,
-  ];
+        airDate,
+        episodeNumber,
+        id,
+        name,
+        overview,
+        productionCode,
+        seasonNumber,
+        showId,
+        stillPath,
+        voteAverage,
+        voteCount,
+      ];
 }
