@@ -31,11 +31,11 @@ class MovieTable extends Equatable {
       );
 
   factory MovieTable.fromDTO(MovieModel movie) => MovieTable(
-    id: movie.id,
-    title: movie.title,
-    posterPath: movie.posterPath,
-    overview: movie.overview,
-  );
+        id: movie.id,
+        title: movie.title,
+        posterPath: movie.posterPath,
+        overview: movie.overview,
+      );
 
   Map<String, dynamic> toJson() => {
         'id': id,
