@@ -7,14 +7,14 @@ class TVResponse extends Equatable {
   const TVResponse({required this.tvList});
 
   factory TVResponse.fromJson(Map<String, dynamic> json) => TVResponse(
-    tvList: List<TVModel>.from((json["results"] as List)
-        .map((x) => TVModel.fromJson(x))
-        .where((element) => element.posterPath != null)),
-  );
+        tvList: List<TVModel>.from((json["results"] as List)
+            .map((x) => TVModel.fromJson(x))
+            .where((element) => element.posterPath != null)),
+      );
 
   Map<String, dynamic> toJson() => {
-    "results": List<dynamic>.from(tvList.map((x) => x.toJson())),
-  };
+        "results": List<dynamic>.from(tvList.map((x) => x.toJson())),
+      };
 
   @override
   List<Object> get props => [tvList];
