@@ -16,39 +16,39 @@ class TVTable extends Equatable {
   });
 
   factory TVTable.fromEntity(TVDetail tv) => TVTable(
-    id: tv.id,
-    name: tv.name,
-    posterPath: tv.posterPath,
-    overview: tv.overview,
-  );
+        id: tv.id,
+        name: tv.name,
+        posterPath: tv.posterPath,
+        overview: tv.overview,
+      );
 
   factory TVTable.fromMap(Map<String, dynamic> map) => TVTable(
-    id: map['id'],
-    name: map['name'],
-    posterPath: map['posterPath'],
-    overview: map['overview'],
-  );
+        id: map['id'],
+        name: map['name'],
+        posterPath: map['posterPath'],
+        overview: map['overview'],
+      );
 
   factory TVTable.fromDTO(TVModel movie) => TVTable(
-    id: movie.id,
-    name: movie.name,
-    posterPath: movie.posterPath,
-    overview: movie.overview,
-  );
+        id: movie.id,
+        name: movie.name,
+        posterPath: movie.posterPath,
+        overview: movie.overview,
+      );
 
   Map<String, dynamic> toJson() => {
-    'id': id,
-    'name': name,
-    'posterPath': posterPath,
-    'overview': overview,
-  };
+        'id': id,
+        'name': name,
+        'posterPath': posterPath,
+        'overview': overview,
+      };
 
   TV toEntity() => TV.watchList(
-    id: id,
-    overview: overview,
-    posterPath: posterPath,
-    name: name,
-  );
+        id: id,
+        overview: overview,
+        posterPath: posterPath,
+        name: name,
+      );
 
   @override
   List<Object?> get props => [id, name, posterPath, overview];
