@@ -57,34 +57,34 @@ class TVDetail extends Equatable {
     required this.type,
     required this.voteAverage,
     required this.voteCount,
-});
+  });
 
   @override
   List<Object?> get props => [
-    adult,
-    backdropPath,
-    episodeRunTime,
-    firstAirDate,
-    genres,
-    homepage,
-    id,
-    inProduction,
-    lastAirDate,
-    lastEpisodeToAir,
-    name,
-    numberOfEpisodes,
-    numberOfSeasons,
-    originCountry,
-    originalLanguage,
-    originalName,
-    overview,
-    popularity,
-    posterPath,
-    seasons,
-    status,
-    tagline,
-    type,
-    voteAverage,
-    voteCount,
-  ];
+        adult,
+        backdropPath,
+        episodeRunTime,
+        firstAirDate,
+        genres,
+        homepage,
+        id,
+        inProduction,
+        lastAirDate,
+        lastEpisodeToAir,
+        name,
+        numberOfEpisodes,
+        numberOfSeasons,
+        originCountry,
+        originalLanguage,
+        originalName,
+        overview,
+        popularity,
+        posterPath,
+        seasons,
+        status,
+        tagline,
+        type,
+        voteAverage,
+        voteCount,
+      ];
 }
