@@ -22,5 +22,5 @@ class DatabaseFailure extends Failure {
 }
 
 class CacheFailure extends Failure {
-  const CacheFailure(String message): super(message);
+  const CacheFailure(String message) : super(message);
 }
