@@ -13,7 +13,7 @@ void main() {
     voteCount: 1,
     firstAirDate: 'firstAirDate',
     name: 'name',
-    originCountry: ['1','2'],
+    originCountry: ['1', '2'],
     originalLanguage: 'originalLanguage',
     originalName: 'originalName',
   );
@@ -24,7 +24,7 @@ void main() {
     id: 1,
     originalName: 'originalName',
     originalLanguage: 'originalLanguage',
-    originCountry: const ['1','2'],
+    originCountry: const ['1', '2'],
     overview: 'overview',
     popularity: 1,
     posterPath: 'posterPath',
@@ -40,7 +40,7 @@ void main() {
     "genre_ids": [1, 2, 3],
     "id": 1,
     "name": 'name',
-    'original_country': ['1','2'],
+    'original_country': ['1', '2'],
     "original_language": 'originalLanguage',
     "original_name": 'originalName',
     "overview": 'overview',
