@@ -18,14 +18,13 @@ void main() {
   const tId = 1;
   final tTVList = <TV>[];
 
-  test('should get list of tv recommendations from the repository',
-          () async {
-        // arrange
-        when(mockTVRepository.getTVRecommendations(tId))
-            .thenAnswer((_) async => Right(tTVList));
-        // act
-        final result = await usecase.execute(tId);
-        // assert
-        expect(result, Right(tTVList));
-      });
+  test('should get list of tv recommendations from the repository', () async {
+    // arrange
+    when(mockTVRepository.getTVRecommendations(tId))
+        .thenAnswer((_) async => Right(tTVList));
+    // act
+    final result = await usecase.execute(tId);
+    // assert
+    expect(result, Right(tTVList));
+  });
 }
