@@ -15,8 +15,7 @@ void main() {
 
   test('should get watchlist status from repository', () async {
     // arrange
-    when(mockTVRepository.isAddedToWatchlist(1))
-        .thenAnswer((_) async => true);
+    when(mockTVRepository.isAddedToWatchlist(1)).thenAnswer((_) async => true);
     // act
     final result = await usecase.execute(1);
     // assert
